requirements.txt
Flask
gunicorn
werkzeug
Pillow
numpy
torch
torchvision
matplotlib
pycocotools
albumentations